Rework code based on review.
Simplify callback passed to patch_with_callback so that it doesn't
get the client object (that's easy enough for it to get if it needs it).

Remove the requirement that the callback returns True/False to
continue/not continue.  Instead, the callback will raise a NoChange
exception to indicate that it doesn't want to make a change after
all.  If the callback doesn't change the patch's old value and doesn't
raise a NoChange exception then an exception will get raised by
patch because it's likely that the callback is not doing something that
it was intended to do.