Avoid the exception handler to throw exception
The handler is calling LOG.exception, which generates a new exception, thus preventing the handler from completing successfully and properly processing the event.
I modified the except block in the handler to print the error message associated with the exception generated in the handler and caught. This should facilitate the troubleshooting in case the handler fails.